requirements.txt
streamlit
langchain

langchain-google-genai
google-generativeai

openai
python-dotenv